% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsae.r
\name{hc}
\alias{hc}
\title{inverse canonical link}
\usage{
hc(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
calculated value
}
\description{
a simple function for inverse canonical link
}
\details{
calculate the inverse canonical link \code{\link{unsae}}
}
\examples{
hc(1)

}
