% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsae.r
\name{multilevel_EM}
\alias{multilevel_EM}
\title{a main function to fit multilevel em with gaussian spatial correlation. Some chunk of the code has been dedicated to handle the syntax. glmer function is called to get the initial value of the estimates.}
\usage{
multilevel_EM(formula, data, tol = 0.001, coordinates)
}
\arguments{
\item{formula}{formula argument following the glmer syntax.}

\item{data}{name of the data set.}

\item{tol}{tolerence level to control the em convergence. default value 1e-3.}

\item{coordinates}{specifying the name of coordinates within the data. These columns *should* exist within the data set provided.}
}
\value{
list of element that are needed for the prediction.
}
\description{
a main function to fit multilevel em with gaussian spatial correlation. Some chunk of the code has been dedicated to handle the syntax. glmer function is called to get the initial value of the estimates.
}
