% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsae.r
\name{predict_em}
\alias{predict_em}
\title{main prediction}
\usage{
predict_em(em_output, test_set)
}
\arguments{
\item{em_output}{the outcome from the multilevel_EM function.}

\item{test_set}{the test set (or new data set). The data set *should* have the predictor columns as well as the coordinate columns, all with the same names with the training data set.}
}
\value{
predicted outcome combining covariates parts Xb and spatial random effects at new data set
}
\description{
a wrapper to calculate the spatial random effect for a new data set.
}
