% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsae.r
\name{spatial_lik}
\alias{spatial_lik}
\title{a simple spatial likelihood
used within multilevel_EM function of \code{\link{unsae}}}
\usage{
spatial_lik(par, D, Y, mu)
}
\arguments{
\item{par}{parameter values}

\item{D}{distance matrix from plgp::distance}

\item{Y}{"observed" response vector}

\item{mu}{current estimate of mu}
}
\value{
negative log likelihood
}
\description{
a simple spatial likelihood
used within multilevel_EM function of \code{\link{unsae}}
}
